+++ nav2_smac_planner/test/test_node2d.cpp
@@ -142,8 +142,10 @@ TEST(Node2DTest, test_node_2d_neighbors)
   unsigned char cost = static_cast<unsigned int>(1);
   nav2_smac_planner::Node2D * node = new nav2_smac_planner::Node2D(1);
   node->setCost(cost);
-  std::function<bool(const unsigned int &, nav2_smac_planner::Node2D * &)> neighborGetter =
-    [&, this](const unsigned int & index, nav2_smac_planner::Node2D * & neighbor_rtn) -> bool
+  std::function<bool(const uint64_t &,
+    nav2_smac_planner::Node2D * &)> neighborGetter =
+    [&, this](const uint64_t & index,
+      nav2_smac_planner::Node2D * & neighbor_rtn) -> bool
     {
       return false;
     };
